+++ src/main/java/org/choongang/guide/controllers/GuideController.java
@@ -1,6 +1,11 @@
 package org.choongang.guide.controllers;
 
 import lombok.RequiredArgsConstructor;
+import org.choongang.admin.board.controllers.RequestBoardPosts;
+import org.choongang.admin.board.entities.NoticeSearch;
+import org.choongang.admin.board.entities.Notice_;
+import org.choongang.admin.board.entities.requestComment;
+import org.choongang.admin.board.service.BoardService;
 import org.choongang.admin.gamecontent.controllers.GameContentSearch;
 import org.choongang.admin.gamecontent.entities.GameContent;
 import org.choongang.admin.gamecontent.service.GameContentInfoService;
@@ -16,6 +21,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Controller
 @RequestMapping("/guide")
@@ -23,6 +29,7 @@ import java.util.List;
 public class GuideController implements ExceptionProcessor  {
 
     private final GameContentInfoService gameContentInfoService;
+    private final BoardService boardService;
 
     @GetMapping
     public String index() {
@@ -53,6 +60,44 @@ public class GuideController implements ExceptionProcessor  {
 
 
         return "guide/product";
+    }
+
+    @GetMapping("/list/noticeFaq")
+    public String adminnoticeFaqList(@ModelAttribute NoticeSearch search, Model model) {
+        commonProcess("notice&faq", model);
+
+        ListData<Notice_> noticeList = boardService.getListOrderByOnTop(search);
+        model.addAttribute("noticeList", noticeList.getItems());
+        model.addAttribute("pagination", noticeList.getPagination());
+
+
+        return "board/noticeFaqList2";
+    }
+
+    @GetMapping("/detail/{num}")
+    public String detail(@PathVariable("num") Long num, @ModelAttribute requestComment form, Model model){
+
+        // 경로 변수 num이 null이거나 음수인 경우에는 board/list/noticeFaq로 리다이렉션
+        if (num <= 0) {
+            return "redirect:/guide/list/noticeFaq";
+        }
+
+        // 게시글 번호를 사용하여 해당 게시글 정보를 가져온다.
+        Optional<Notice_> noticeDetail = boardService.findByNum(num);
+
+        // 게시글이 존재하는 경우에는 모델에 추가하고 board/noitceDetail2 페이지를 반환
+        if (noticeDetail.isPresent()) {
+
+            boardService.increaseVisitCount(num);
+            model.addAttribute("noticeDetail", noticeDetail.get());
+            model.addAttribute("requestBoardPosts", new RequestBoardPosts());
+
+            return "board/noticeFaqDetail2";
+        }
+
+        // 해당 게시글을 찾을 수 없는 경우에는 board/list/noticeFaq로 리다이렉션
+        return "redirect:/guide/list/noticeFaq";
+
     }
 
     private void commonProcess(String mode, Model model) {
